refactor: 클래스 패키지 분리 및 필드 final로 변경
