+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,13 +1,10 @@
-package roomescape;
+package roomescape.controller;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,13 +13,15 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequest;
 
 @RestController
 @RequestMapping("reservations")
 public class ReservationController {
 
-    private Map<Long, Reservation> reservations = new HashMap<>();
-    private AtomicLong index = new AtomicLong();
+    private final Map<Long, Reservation> reservations = new HashMap<>();
+    private final AtomicLong index = new AtomicLong();
 
     @GetMapping
     public List<Reservation> findAll() {
