+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,14 +1,14 @@
-package roomescape;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
 public class Reservation {
 
-    private Long id;
-    private String name;
-    private LocalDate date;
-    private LocalTime time;
+    private final Long id;
+    private final String name;
+    private final LocalDate date;
+    private final LocalTime time;
 
     public Reservation(Long id, String name, LocalDate date, LocalTime time) {
         this.id = id;
